requirements.txt
# Project dependencies
# Add your required packages here
requests==2.32.4
mysql-connector-python==8.0.33
pymongo==4.14.0
sqlalchemy==2.0.43
fastapi==0.116.1
uvicorn==0.35.0
openai==1.59.3